chore: create supabaseStorageClient
