+++ src/lib/storage/index.ts
@@ -1,2 +1,3 @@
+export * from './StorageApi'
 export * from './StorageClient'
 export * from './types'
